mv editor api typings
